Use _requirestxt logic for writing 'requires.txt'
This change extracts the logic for writing 'requires.txt' from
the ``egg_info`` module into ``_requirestxt``.

It also adds the missing pre-processing for ``install_requires`` and
``extras_require`` previously done in ``dist`` but now implemented in
``_requires.txt``.
